--- a/lib/rust/profiler/macros/src/lib.rs
+++ b/lib/rust/profiler/macros/src/lib.rs
@@ -60,8 +60,13 @@ fn define_profiler(
             // === Trait Implementations ===
 
             impl Profiler for #obj_ident {
-                fn start(parent: EventId, label: Label, time: Option<Timestamp>) -> Self {
-                    #obj_ident(EventLog.start(parent, label, time))
+                fn start(
+                    parent: EventId,
+                    label: StaticLabel,
+                    time: Option<Timestamp>,
+                    start: StartState,
+                ) -> Self {
+                    #obj_ident(EventLog.start(parent, label, time, start))
                 }
                 fn finish(self) {
                     EventLog.end(self.0, Timestamp::now())
@@ -82,7 +87,7 @@ fn define_profiler(
             macro_rules! #start {
                 ($parent: expr, $label: expr) => {{
                     use profiler::Parent;
-                    let label = concat!($label, " (", file!(), ":", line!(), ")");
+                    let label = profiler::Label(concat!($label, " (", file!(), ":", line!(), ")"));
                     let profiler: profiler::Started<profiler::#obj_ident> = $parent.new_child(label);
                     profiler
                 }}
@@ -93,7 +98,7 @@ fn define_profiler(
             macro_rules! #with_same_start {
                 ($parent: expr, $label: expr) => {{
                     use profiler::Parent;
-                    let label = concat!($label, " (", file!(), ":", line!(), ")");
+                    let label = profiler::Label(concat!($label, " (", file!(), ":", line!(), ")"));
                     let profiler: profiler::Started<profiler::#obj_ident> =
                         $parent.new_child_same_start(label);
                     profiler
@@ -114,7 +119,14 @@ fn define_profiler(
             // === Trait Implementations ===
 
             impl Profiler for #obj_ident {
-                fn start(_: EventId, _: Label, _: Option<Timestamp>) -> Self { Self(()) }
+                fn start(
+                    _: EventId,
+                    _: StaticLabel,
+                    _: Option<Timestamp>,
+                    _: StartState,
+                ) -> Self {
+                    Self(())
+                }
                 fn finish(self) {}
                 fn pause(&self) {}
                 fn resume(&self) { }
@@ -150,12 +162,12 @@ fn enabled_impl_parent(
 ) -> proc_macro::TokenStream {
     let ts = quote::quote! {
         impl Parent<#child_ident> for #parent_ident {
-            fn new_child(&self, label:Label) -> Started<#child_ident> {
+            fn new_child(&self, label: StaticLabel) -> Started<#child_ident> {
                 let start = Some(Timestamp::now());
-                Started(#child_ident::start(self.0, label, start))
+                Started(#child_ident::start(self.0, label, start, StartState::Active))
             }
-            fn new_child_same_start(&self, label:Label) -> Started<#child_ident> {
-                Started(#child_ident::start(self.0, label, None))
+            fn new_child_same_start(&self, label: StaticLabel) -> Started<#child_ident> {
+                Started(#child_ident::start(self.0, label, None, StartState::Active))
             }
         }
     };
@@ -170,10 +182,10 @@ fn disabled_impl_parent(
 ) -> proc_macro::TokenStream {
     let ts = quote::quote! {
         impl Parent<#child_ident> for #parent_ident {
-            fn new_child(&self, label: Label) -> Started<#child_ident> {
+            fn new_child(&self, label: StaticLabel) -> Started<#child_ident> {
                 self.new_child_same_start(label)
             }
-            fn new_child_same_start(&self, _label: Label) -> Started<#child_ident> {
+            fn new_child_same_start(&self, _label: StaticLabel) -> Started<#child_ident> {
                 Started(#child_ident(()))
             }
         }
@@ -295,17 +307,17 @@ fn start_profiler(
     label: String,
     asyncness: bool,
 ) -> proc_macro2::TokenStream {
-    let start_await = match asyncness {
-        true => quote::quote! { profiler.pause(); },
-        false => quote::quote! {},
+    let state = match asyncness {
+        true => quote::quote! { profiler::StartState::Paused },
+        false => quote::quote! { profiler::StartState::Active },
     };
     quote::quote! {
         let __profiler_scope = {
             use profiler::Profiler;
-            let parent = profiler::IMPLICIT_ID;
+            let parent = profiler::EventId::implicit();
             let now = Some(profiler::Timestamp::now());
-            let profiler = profiler::#obj_ident::start(parent, #label, now);
-            #start_await
+            let label = profiler::Label(#label);
+            let profiler = profiler::#obj_ident::start(parent, label, now, #state);
             profiler::Started(profiler)
         };
     }