@kazcw 2022-02-24
@MichaelMauderer While doing a high-level review of #3297, I realized data consumer should have an easier way to distinguish async from non-async Measurements. I'll revise that now...