@kazcw 2022-02-24
A `Measurement` now has a `Lifetime` enum that can be used to handle sync/async measurements differently; async measurements have a `create_to_finish` convenience method that we can use when graphing to show a profiler's total extent.