@kazcw 2022-02-25
@wdanilo Re: performance

I don't expect any performance change. Benchmarks are the same within variance.

#### BEFORE

test bench::log_measurements_1000   ... bench:       3,892 ns/iter (+/- 581)
test bench::log_measurements_10_000 ... bench:      36,740 ns/iter (+/- 290)

#### AFTER

test bench::log_measurements_1000   ... bench:       3,679 ns/iter (+/- 79)
test bench::log_measurements_10_000 ... bench:      37,063 ns/iter (+/- 34,044)
test bench::push_vec_1000           ... bench:       1,997 ns/iter (+/- 74)
test bench::push_vec_10_000         ... bench:      20,100 ns/iter (+/- 651)