@kazcw 2022-02-25
### Modules

#### `profiler`

Moved implementation detalis into `profiler::internal`; now the top-level contents of `profiler` contain only the API intended for general use (and described in the design doc). `profiler::internal` is used to implement the macros, and used by `profiler::data` to decode raw event logs.

#### `profiler::data`

Moved parsing implementation into `profiler::data::parse`. The top-level module now only defines the interface.