@wdanilo 2022-02-25
@kazcw please never do this:
<img width="799" alt="image" src="https://user-images.githubusercontent.com/1623053/155800704-c6a3f89c-a0bd-42f7-b4ca-c9fc99a5b5a8.png">

Now, when browsing the code, I cant see old discussions and my old review comments.